+++ includes/class-gc-testimonials-to-testimonials.php
@@ -13,13 +13,7 @@
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-if ( defined( 'TW_DIR_LIB' ) ) {
-	$plugin_base = TW_DIR_LIB;
-} else {
-	$plugin_base = WP_PLUGIN_DIR . '/testimonials-widget/includes/libraries/';
-}
-
-require_once $plugin_base . 'aihrus-framework/class-aihrus-common.php';
+require_once GCT2T_DIR_LIB_ALT . 'aihrus-framework/class-aihrus-common.php';
 require_once GCT2T_DIR_INC . 'class-gc-testimonials-to-testimonials-settings.php';
 
 if ( class_exists( 'Gc_Testimonials_to_Testimonials' ) )
@@ -600,21 +594,6 @@ class Gc_Testimonials_to_Testimonials extends Aihrus_Common {
 
 	public static function version_check() {
 		$valid_version = true;
-
-		$valid_base = true;
-		if ( ! is_plugin_active( GCT2T_REQ_BASE ) ) {
-			$valid_base = false;
-		} elseif ( ! defined( 'TW_VERSION' ) ) {
-			$valid_base = false;
-		} elseif ( ! version_compare( TW_VERSION, GCT2T_REQ_VERSION, '>=' ) ) {
-			$valid_base = false;
-		}
-
-		if ( ! $valid_base ) {
-			$valid_version = false;
-			self::set_notice( 'gct2t_notice_version' );
-		}
-
 		if ( ! $valid_version ) {
 			deactivate_plugins( self::BASE );
 			self::check_notices();